# Don't split multi-line rows across batches