--- a/internal/batch/scan.go
+++ b/internal/batch/scan.go
@@ -12,16 +12,23 @@ type Options struct {
 	Size  int   // maximum number of rows per batch
 	Skip  int   // how many header lines to skip at the beginning
 	Limit int64 // total number of rows to scan after the header
+
+	Quote  byte // the QUOTE character; defaults to '"'
+	Escape byte // the ESCAPE character; defaults to QUOTE
 }
 
 // Scan reads all lines from an io.Reader, partitions them into net.Buffers with
-// opts.Size lines each, and writes each batch to the out channel. If opts.Skip
+// opts.Size rows each, and writes each batch to the out channel. If opts.Skip
 // is greater than zero, that number of lines will be discarded from the
 // beginning of the data. If opts.Limit is greater than zero, then Scan will
-// stop once it has written that number of lines, across all batches, to the
+// stop once it has written that number of rows, across all batches, to the
 // channel.
+//
+// Scan expects the input to be in Postgres CSV format. Since this format allows
+// rows to be split over multiple lines, the caller may provide opts.Quote and
+// opts.Escape as the QUOTE and ESCAPE characters used for the CSV input.
 func Scan(r io.Reader, out chan<- net.Buffers, opts Options) error {
-	var linesRead int64
+	var rowsRead int64
 	reader := bufio.NewReader(r)
 
 	for skip := opts.Skip; skip > 0; {
@@ -37,10 +44,22 @@ func Scan(r io.Reader, out chan<- net.Buffers, opts Options) error {
 		}
 
 		if !isPrefix {
-			// We pulled a full line from the buffer.
+			// We pulled a full row from the buffer.
 			skip--
 		}
 	}
+
+	quote := byte('"')
+	if opts.Quote != 0 {
+		quote = opts.Quote
+	}
+
+	escape := quote
+	if opts.Escape != 0 {
+		escape = opts.Escape
+	}
+
+	scanner := makeCSVRowState(quote, escape)
 
 	// We read a continuous stream of []byte from our buffered reader. Rather
 	// than coalesce all of the incoming slices into a single contiguous buffer
@@ -50,9 +69,11 @@ func Scan(r io.Reader, out chan<- net.Buffers, opts Options) error {
 	// the slices as-is and store them in net.Buffers, which is a convenient
 	// io.Reader abstraction wrapped over a [][]byte.
 	bufs := make(net.Buffers, 0, opts.Size)
-	var bufferedLines int
+	var bufferedRows int
 
 	for {
+		eol := false
+
 		data, err := reader.ReadSlice('\n')
 
 		switch err {
@@ -67,8 +88,7 @@ func Scan(r io.Reader, out chan<- net.Buffers, opts Options) error {
 
 		case nil:
 			// We read a full line from the input.
-			bufferedLines++
-			linesRead++
+			eol = true
 
 		default:
 			return err
@@ -81,17 +101,25 @@ func Scan(r io.Reader, out chan<- net.Buffers, opts Options) error {
 			copy(buf, data)
 			bufs = append(bufs, buf)
 
-			if bufferedLines >= opts.Size { // dispatch to COPY worker & reset
+			// Figure out whether we're still inside a quoted value, in which
+			// case the row hasn't ended yet even if we're at the end of a line.
+			scanner.Scan(buf)
+			if eol && !scanner.NeedsMore() {
+				bufferedRows++
+				rowsRead++
+			}
+
+			if bufferedRows >= opts.Size { // dispatch to COPY worker & reset
 				out <- bufs
 				bufs = make(net.Buffers, 0, opts.Size)
-				bufferedLines = 0
+				bufferedRows = 0
 			}
 		}
 
 		// Check termination conditions.
 		if err == io.EOF {
 			break
-		} else if opts.Limit != 0 && linesRead >= opts.Limit {
+		} else if opts.Limit != 0 && rowsRead >= opts.Limit {
 			break
 		}
 	}
@@ -102,4 +130,96 @@ func Scan(r io.Reader, out chan<- net.Buffers, opts Options) error {
 	}
 
 	return nil
+}
+
+// csvRowState maintains a basic parse state for the Postgres CSV format. It's
+// designed to help clients maintain an accurate row count even when quoted rows
+// are split over multiple lines.
+type csvRowState struct {
+	quote, escape byte // delimiters
+
+	// scan state
+	inQuote  bool // are we in an open quoted value?
+	inEscape bool // are we (potentially) in an open escape sequence?
+}
+
+// makeCSVRowState initializes a new csvRowState with the given delimiters.
+func makeCSVRowState(quote, escape byte) *csvRowState {
+	return &csvRowState{
+		quote:  quote,
+		escape: escape,
+	}
+}
+
+// Scan reads raw line data from a CSV stream. After feeding data to Scan, the
+// parse state may be checked by calling NeedsMore.
+//
+// Note that newline characters (and whitespace characters in general) are
+// significant in Postgres CSV format and MUST NOT be stripped from the stream
+// that is given to Scan. (This property is met by bufio.Reader's ReadSlice
+// method.)
+func (c *csvRowState) Scan(buf []byte) {
+	for _, b := range buf {
+		// If we think the previous character might have been an escape, the
+		// current character might need to be ignored.
+		if c.inEscape {
+			c.inEscape = false
+
+			// Only the quote and escape delimiters can themselves be escaped.
+			switch b {
+			case c.quote, c.escape:
+				// Okay, it was really an escape and we should ignore this.
+				continue
+			}
+
+			// NB: this is the strangest corner case of the Postgres CSV
+			// format. If the quote and escape delimiters are the same character
+			// -- e.g. in the default case, where both are a double-quote (") --
+			// and we're to this point in the code, then the last escape
+			// character we saw was actually an ending quote, and we need to
+			// make sure the state reflects that before continuing.
+			//
+			// (As a concrete example, consider a buffer with the following
+			// data at the start:
+			//
+			//     "hello world"
+			//
+			// If the next character in the buffer is a double-quote, then we're
+			// still inside an open quoted value, since the "" sequence is
+			// replaced with a literal double-quote. But if the next character
+			// is a comma, then this is a complete quoted value.)
+			//
+			// The Postgres code handles this case with a lookahead, but we
+			// don't always have the ability to do that here, since we're
+			// operating on a buffer stream.
+			if c.quote == c.escape {
+				c.inQuote = false
+			}
+		}
+
+		// Escape sequences are only recognized inside of a quoted value;
+		// otherwise the escape character has no special meaning.
+		if c.inQuote && (b == c.escape) {
+			c.inEscape = true
+			continue
+		}
+
+		if b == c.quote {
+			// We know this is an unescaped quote delimiter, so we're either
+			// beginning or ending a quoted string.
+			c.inQuote = !c.inQuote
+		}
+	}
+}
+
+// NeedsMore returns true if the current row is incomplete: the previous buffers
+// given to Scan opened a quoted value that has not yet been closed.
+//
+// Note that even if NeedsMore returns false, that does NOT imply that the
+// current position is at the end of a row. (To decide that, the caller needs to
+// track whether the stream is also at the end of a line.)
+func (c *csvRowState) NeedsMore() bool {
+	// We don't need to check c.inEscape, because that can only be true if
+	// c.inQuote is also true.
+	return c.inQuote
 }