@jchampio 2022-08-11
So I've done several attempts at a SIMD optimization for this. `bytes.IndexByte()` would be far faster for the case where there are only a few quotes scattered throughout, but if quotes are common (for example, the UK Price Paid Data CSV quotes every column) then it makes things _worse_. (The Go internals use a tuning factor to try to switch back to brute force optimization if they detect that SIMD is slowing things down.)

My current IndexByte-based attempt works, but the code is awful, and I'm hoping somebody has some better ideas while I clean it up for review.