@svenklemm 2022-08-15
Can you elaborate what your benchmark numbers mean. What is op?