@jchampio 2022-08-15
@svenklemm `op` here refers to one iteration of the [`Scan` benchmark loop](https://github.com/timescale/timescaledb-parallel-copy/pull/67/commits/d3a405f296ed950a62e590ab6e7f4de77c3e5cf2#diff-6c1ccd8aa924b4075404c8ff2800b89f3647575dbb2e647390da38f33abffc4fR308-R315).

[`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) determines whether two sets of benchmark results are statistically significantly different. So for the line
```
name                                                    old time/op  new time/op  delta
...
Scan/no_quotes_(standard_escapes)-8                      273µs ± 2%   625µs ± 2%  +129.08%  (p=0.008 n=5+5)
```
this means that for the `no_quotes_(standard_escapes)` test, the old version of `Scan` took an average of 273 microseconds per benchmark iteration (with the min and max durations within 2% of the average), and the new version took 625 microseconds on average, for a 129% slowdown.

The p-value of the [U-test](https://en.wikipedia.org/wiki/Mann%E2%80%93Whitney_U_test) is 0.008, meaning that there would be a 0.8% chance of getting a rank-sum result of this significance due to completely random chance. (This is probably not particularly meaningful for the small `n` values I'm testing -- `n=5+5` here means that there are 5 samples from the old and 5 from the new. But the two spreads don't overlap at all, hinting pretty strongly at a significant difference, so I'm not too worried about chasing phantoms. :grin: We can switch to a [two-sample t-test](https://en.wikipedia.org/wiki/Welch%27s_t-test) if desired.)