@svenklemm 2022-08-16
I meant what specific operation does op refer to? Is it parsing a single line, parsing 5 lines, something else?