@jchampio 2022-08-16
In this case, parsing 500 lines (see the setup for `BenchmarkScan`).