@svenklemm 2022-08-18
OK so if op is 5k lines that would put us at theoretical limit of 5M to 10M rows per seconds. How much time is spent in scanning vs the rest of the code during an actual run?