@jchampio 2022-08-18
>OK so if op is 5k lines that would put us at theoretical limit of 5M to 10M rows per seconds.

On my dev machine, yes. I'm still working on grabbing an AWS instance so we can see what it's like on something more throttled.

> How much time is spent in scanning vs the rest of the code during an actual run?

With this patch, Scan takes up roughly 70% of the CPU time for a [`yellow_tripdata`](https://www1.nyc.gov/site/tlc/about/tlc-trip-record-data.page) ingestion.