@svenklemm 2022-08-19
2 million will probably be enough to saturate any singlenode instance so i think on that side we would still have good buffer.