@jchampio 2022-08-24
For posterity, I ran the benchmark on an EC2 `t2.micro` instance (I figured the free tier was probably a relevant case for users :grin:).
```
goos: linux
goarch: amd64
pkg: github.com/timescale/timescaledb-parallel-copy/internal/batch
cpu: Intel(R) Xeon(R) CPU E5-2676 v3 @ 2.40GHz
```
benchstat reports
```
name                                                  old time/op  new time/op  delta
Scan/no_quotes_(standard_escapes)                      756µs ± 6%  1555µs ± 6%  +105.76%  (p=0.008 n=5+5)
Scan/no_quotes_(custom_escapes)                        756µs ± 6%  1544µs ± 3%  +104.40%  (p=0.008 n=5+5)
Scan/some_quotes_at_the_beginning_(standard_escapes)   778µs ± 5%  1815µs ± 3%  +133.40%  (p=0.008 n=5+5)
Scan/some_quotes_at_the_beginning_(custom_escapes)     778µs ± 5%  1764µs ± 1%  +126.87%  (p=0.008 n=5+5)
Scan/some_quotes_in_the_middle_(standard_escapes)      777µs ± 3%  1944µs ± 4%  +150.35%  (p=0.008 n=5+5)
Scan/some_quotes_in_the_middle_(custom_escapes)        777µs ± 3%  1940µs ± 4%  +149.78%  (p=0.008 n=5+5)
Scan/all_quotes_(standard_escapes)                     811µs ± 6%  2405µs ± 3%  +196.53%  (p=0.008 n=5+5)
Scan/all_quotes_(custom_escapes)                       811µs ± 6%  2424µs ± 4%  +198.90%  (p=0.008 n=5+5)
Scan/nothing_but_quotes_(standard_escapes)             739µs ± 3%  1761µs ± 7%  +138.30%  (p=0.008 n=5+5)
Scan/nothing_but_quotes_(custom_escapes)               739µs ± 3%  2232µs ± 1%  +202.08%  (p=0.008 n=5+5)
```
So again a 2x-3x slowdown. Previous performance was roughly 6M rows/s, and this change takes us right down to the 2M rows/s threshold you mentioned. So there's no buffer left, but since this is fairly worst-case (throttled cloud CPU) that seems pretty reasonable.